+++ R/math.R
@@ -25,50 +25,49 @@
 Math.errors <- function(x, ...) {
   switch(
     .Generic,
-    "abs" = NextMethod(),
-    "sign" = as.numeric(NextMethod()),
+    "abs" = set_errors(unclass(NextMethod()), errors(x)),
+    "sign" = drop_errors(NextMethod()),
     "sqrt" = x^set_errors(0.5),
     "floor" = , "ceiling" = , "trunc" = , "round" = , "signif" = {
-      values <- .v(NextMethod())
-      e <- errors(x) + abs(.v(x) - values)
-      structure(values, "errors" = e, class = "errors")
+      xx <- .v(NextMethod())
+      set_errors(xx, errors(x) + abs(.v(x) - xx))
     },
     {
-      e <- switch(
+      deriv <- switch(
         .Generic,
-        "exp" = abs(errors(x) * exp(.v(x))),
-        "log" = abs(errors(x) / .v(x) / log(if (missing(...)) exp(1) else c(...)[1])),
-        "expm1" = abs(errors(x) * exp(.v(x))),
-        "log1p" = abs(errors(set_errors(1)+x) / .v(set_errors(1)+x)),
-        "cos" = abs(errors(x) * sin(.v(x))),
-        "sin" = abs(errors(x) * cos(.v(x))),
-        "tan" = errors(x) / cos(.v(x))^2,
-        "cospi" = abs(errors(x) * pi * sin(pi * .v(x))),
-        "sinpi" = abs(errors(x) * pi * cos(pi * .v(x))),
-        "tanpi" = errors(x) * pi / cos(pi * .v(x))^2,
-        "acos" = , "asin" = errors(x) / sqrt(1 - .v(x)^2),
-        "atan" = errors(x) / (1 + .v(x)^2),
-        "cosh" = abs(errors(x) * sinh(.v(x))),
-        "sinh" = abs(errors(x) * cosh(.v(x))),
-        "tanh" = errors(x) / cosh(.v(x))^2,
-        "acosh" = errors(x) / sqrt(.v(x) - 1) / sqrt(.v(x) + 1),
-        "asinh" = errors(x) / sqrt(1 + .v(x)^2),
-        "atanh" = abs(errors(x) / (1 - .v(x)^2)),
-        "cumsum" = propagate(cummatrix(errors(x)))
+        "exp" = , "expm1" = exp(.v(x)),
+        "log" = 1 / .v(x) / log(if (missing(...)) exp(1) else c(...)[1]),
+        "log1p" = 1 / .v(set_errors(1)+x),
+        "cos" = -sin(.v(x)),
+        "sin" = cos(.v(x)),
+        "tan" = 1 / cos(.v(x))^2,
+        "cospi" = -pi * sin(pi * .v(x)),
+        "sinpi" = pi * cos(pi * .v(x)),
+        "tanpi" = pi / cos(pi * .v(x))^2,
+        "acos" = -1 / sqrt(1 - .v(x)^2),
+        "asin" = 1 / sqrt(1 - .v(x)^2),
+        "atan" = 1 / (1 + .v(x)^2),
+        "cosh" = sinh(.v(x)),
+        "sinh" = cosh(.v(x)),
+        "tanh" = 1 / cosh(.v(x))^2,
+        "acosh" = 1 / sqrt(.v(x) - 1) / sqrt(.v(x) + 1),
+        "asinh" = 1 / sqrt(1 + .v(x)^2),
+        "atanh" = 1 / (1 - .v(x)^2)
       )
-      structure(NextMethod(), "errors" = e, class = "errors")
+      propagate(unclass(NextMethod()), x, NA, deriv, NA)
     },
+    "cumsum" = set_errors(unclass(NextMethod()), sqrt(colSums(cummatrix(errors(x))^2))),
     "cumprod" = {
-      values <- NextMethod()
-      e <- propagate(cummatrix(errors(x)) * t(values / t(cummatrix(.v(x), fill=1))))
-      structure(values, "errors" = e, class = "errors")
+      xx <- NextMethod()
+      e <- sqrt(colSums((cummatrix(errors(x)) * t(xx / t(cummatrix(.v(x), fill=1))))^2))
+      set_errors(unclass(xx), e)
     },
     "cummax" = , "cummin" = {
-      values <- NextMethod()
-      indexes <- which(values == .v(x))
+      xx <- NextMethod()
+      indexes <- which(xx == .v(x))
       reps <- diff(c(indexes, length(x)+1))
       e <- rep(errors(x)[indexes], times=reps)
-      structure(values, "errors" = e, class = "errors")
+      set_errors(unclass(xx), e)
     },
     "lgamma" = , "gamma" = , "digamma" = , "trigamma" =
       stop("method '", .Generic, "' not supported for 'errors' objects")
@@ -87,5 +86,5 @@ log2.errors <- function(x) log(x, 2)
 # atan2.errors <- function(x, y) {
 #   z <- y/x
 #   e <- errors(z) / (1 + .v(z)^2)
-#   structure(NextMethod(), "errors" = e, class = "errors")
+#   set_errors(unclass(NextMethod()), e)
 # }