+++ R/ops.R
@@ -39,23 +39,19 @@ Ops.errors <- function(e1, e2) {
     }
   }
 
-  e <- switch(
+  deriv <- switch(
     .Generic,
-    `+` = , `-` =
-      if (!missing(e2))
-        propagate(errors(e1), errors(e2))
-      else errors(e1),
-    `*` =
-      propagate(errors(e1) * .v(e2),
-                errors(e2) * .v(e1)),
-    `/` = , `%%` = , `%/%` =
-      propagate(errors(e1) / .v(e2),
-                errors(e2) * .v(e1) / .v(e2)^2),
-    `^` =
-      propagate(errors(e1) * .v(e1)^(.v(e2)-1) * .v(e2),
-                errors(e2) * .v(e1)^.v(e2) * log(abs(.v(e1))))
+    "+" = , "-" =
+      if (missing(e2)) {
+        e2 <- NA
+        list(do.call(.Generic, list(1)), NA)
+      }
+      else list(1, do.call(.Generic, list(1))),
+    "*" = list(.v(e2), .v(e1)),
+    "/" = , "%%" = , "%/%" = list(1 / .v(e2), -.v(e1) / .v(e2)^2),
+    "^" = list(.v(e1)^(.v(e2)-1) * .v(e2), .v(e1)^.v(e2) * log(abs(.v(e1))))
   )
-  structure(NextMethod(), "errors" = e, class = "errors")
+  propagate(unclass(NextMethod()), e1, e2, deriv[[1]], deriv[[2]])
 }
 
 #' #' Matrix Multiplication