+++ tests/testthat/test-math.R
@@ -21,52 +21,34 @@ test_that("rounding methods work properly", {
 })
 
 test_that("math methods work properly", {
-  Dx <- function(expr) {
-    if (identical(expr, expression(asinh(x))))
-      expression(1 / sqrt(1 + x^2))
-    else if (identical(expr, expression(acosh(x))))
-      expression(1 / (sqrt(x - 1) * sqrt(x + 1)))
-    else if (identical(expr, expression(atanh(x))))
-      expression(1 / (1 - x^2))
-    else
-      D(expr, "x")
-  }
-
-  test_expr <- function(expr) {
-    res <- eval(expr)
-    err <- errors(x)
-    x <- as.numeric(x)
-    expect_equal(res, set_errors(res, abs(err * eval(Dx(expr)))))
-  }
-
   xval <- 1.1:10.1/100
   xerr <- seq(0.005, 0.05, 0.005)/100
   x <- set_errors(xval, xerr)
 
-  test_expr(expression(sqrt(x)))
-  test_expr(expression(exp(x)))
-  test_expr(expression(log(x)))
+  test_expr(sqrt(x))
+  test_expr(exp(x))
+  test_expr(log(x))
   expect_equal(log(x, 10), log10(x))
   expect_equal(log(x, 2), log2(x))
-  expect_equal(log(1 + x), log1p(x))
-  expect_equal(exp(x) - 1, expm1(x))
-  test_expr(expression(sin(x)))
-  test_expr(expression(cos(x)))
-  test_expr(expression(tan(x)))
-  test_expr(expression(asin(x)))
-  test_expr(expression(acos(x)))
-  test_expr(expression(atan(x)))
-  expect_equal(sin(pi*x), sinpi(x))
-  expect_equal(cos(pi*x), cospi(x))
-  expect_equal(tan(pi*x), tanpi(x))
-  test_expr(expression(sinh(x)))
-  test_expr(expression(cosh(x)))
-  test_expr(expression(tanh(x)))
-  test_expr(expression(asinh(x)))
-  x <- x*100
-  test_expr(expression(acosh(x)))
-  x <- x/100
-  test_expr(expression(atanh(x)))
+  expect_equal(log(set_errors(1) + x), log1p(x))
+  expect_equal(exp(x) - set_errors(1), expm1(x))
+  test_expr(sin(x))
+  test_expr(cos(x))
+  test_expr(tan(x))
+  test_expr(asin(x))
+  test_expr(acos(x))
+  test_expr(atan(x))
+  expect_equal(sin(set_errors(pi)*x), sinpi(x))
+  expect_equal(cos(set_errors(pi)*x), cospi(x))
+  expect_equal(tan(set_errors(pi)*x), tanpi(x))
+  test_expr(sinh(x))
+  test_expr(cosh(x))
+  test_expr(tanh(x))
+  test_expr(asinh(x))
+  x <- x * set_errors(100)
+  test_expr(acosh(x))
+  x <- x / set_errors(100)
+  test_expr(atanh(x))
 })
 
 test_that("cumulative methods work properly", {